chore: validate api users로 이동
