+++ src/api/users/index.ts
@@ -1,6 +1,8 @@
 import { kyFileInstance, kyInstance, kyPublicInstance } from "@/api";
 import type { GroupListResponse } from "@/api/groups/type";
 import type { UserResponse } from "@/api/users/type";
+import { HTTP_ERROR_STATUS } from "@/shared/constant/api";
+import { HTTPError } from "ky";
 
 export const getGroupsByUsers = async (userNickname: string) => {
   const response = await kyInstance
@@ -56,4 +58,36 @@ export const postSignUp = async (formData: FormData) => {
   });
 
   return response;
+};
+
+export const validateNickname = async (nickname: string) => {
+  try {
+    const response = await checkNickname(nickname);
+
+    return response;
+  } catch (err) {
+    if (err instanceof HTTPError) {
+      const { response } = err;
+
+      if (response.status === HTTP_ERROR_STATUS.CONFLICT) {
+        return null;
+      }
+    }
+  }
+};
+
+export const validateBojNickname = async (nickname: string) => {
+  try {
+    const response = await checkBojNickname(nickname);
+
+    return response;
+  } catch (err) {
+    if (err instanceof HTTPError) {
+      const { response } = err;
+
+      if (response.status === HTTP_ERROR_STATUS.NOT_FOUND) {
+        return null;
+      }
+    }
+  }
 };
